requirements.txt
django==4.0.1
django-crispy-forms==1.14.0
xhtml2pdf==0.2.5
Pillow==9.0.0
django-import-export==2.7.1
psycopg2-binary==2.9.3
django_filter==21.1
django-bootstrap-form==3.4
django-widget-tweaks==1.4.12
django-bstrap-modals==2.4.0
reportlab==3.6.7
django-encrypted-model-fields==0.6.1